backend/requirements.txt
yfinance==0.2.50
tableau-api-lib==0.1.50
python-dotenv==1.0.1
supabase==2.10.0
pandas==2.2.3
flask==3.1.0